requirements-win.txt
rich==13.7.0
requests==2.31.0
firebase-admin==6.4.0
python-dateutil==2.8.2
cryptography>=41.0.5
packaging>=21.0
tqdm==4.66.2
argparse==1.4.0
beautifulsoup4==4.12.3
# Using a pre-built wheel for lxml
lxml>=5.1.0
pyjwt==2.8.0
colorama==0.4.6
tabulate==0.9.0
pyyaml==6.0.1
pillow==10.2.0

# Security-focused packages
bandit==1.7.7
safety>=2.3.5
python-nmap==0.7.1
dnspython==2.6.1
pyOpenSSL>=24.0.0
sslyze>=5.2.0
python-whois==0.8.0

# Development dependencies
pytest==8.0.2
black==24.2.0
mypy==1.8.0
flake8==7.0.0
isort==5.13.2 